queue/requirements.txt
paho-mqtt
boto3